feat: move __tests__ and __mocks__ under test/unit folder
